+++ src/app.html
@@ -6,7 +6,17 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <meta name="theme-color" content="#333333">
 
-    <link rel="stylesheet" href="global.css">
+    <style>
+        body {
+            margin: 0;
+            padding: 0;
+            font-family: Courier, monospace;
+        }
+        
+        * {
+            box-sizing: border-box;
+        }
+    </style>
     <link rel="manifest" href="manifest.json" crossorigin="use-credentials">
     <link rel="icon" type="image/gif" href="text.gif">
