+++ src/routes/[...slug].svelte
@@ -5,7 +5,7 @@
   export async function load({ page, fetch }) {
     // the `slug` parameter is available because
     // this file is called [slug].svelte
-    const res = await fetch(`${page.params.slug}.json`);
+    const res = await fetch(`/pieces/${page.params.slug}`);
     const data = await res.json();
 
     if (res.status === 200) {
