+++ src/routes/__error.svelte
@@ -2,7 +2,7 @@
 	export let status: number;
 	export let error: Error;
 
-	const dev = process.env.NODE_ENV === 'development';
+	const dev = false;
 </script>
 
 <style>
