--- static/global.css
@@ -1,9 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: Courier, monospace;
-}
-
-* {
-  box-sizing: border-box;
-}
